DOMA-1803 BaseMeterModalForm, update and delete meter
