+++ apps/condo/domains/meter/utils/clientSchema/search.ts
@@ -1,3 +1,4 @@
+import { isEmpty } from 'lodash'
 import { gql } from 'graphql-tag'
 
 const GET_METER_QUERY = gql`
@@ -60,4 +61,25 @@ export async function searchMeterResource (client, where, orderBy, first = 10, s
     if (error) console.warn(error)
     if (data) return data.objs.map(x => ({ text: x.address, value: x.id }))
     return []
+}
+
+const GET_ALL_METER_RESOURCES_BY_VALUE_QUERY = gql`
+    query selectMeterResources ($where: MeterResourceWhereInput, $orderBy: String, $first: Int, $skip: Int) {
+        objs: allMeterResources(where: $where, orderBy: $orderBy, first: $first, skip: $skip) {
+            id
+            name
+        }
+    }
+`
+
+export async function searchMeterResources (client, searchText, query = {}, first = 10, skip = 0) {
+    const where = {
+        ...!isEmpty(searchText) ? { name_contains_i: searchText } : {},
+        ...query,
+    }
+    const orderBy = 'name_ASC'
+    const { data = [], error } = await _search(client, GET_ALL_METER_RESOURCES_BY_VALUE_QUERY, { where, orderBy, first, skip })
+    if (error) console.warn(error)
+
+    return data.objs.map(({ name, id }) => ({ text: name, value: id }))
 }
