+++ apps/condo/lang/en.json
@@ -359,7 +359,7 @@
   "pages.condo.meter.Place": "Place",
   "pages.condo.meter.MeterReading": "Reading",
   "pages.condo.meter.MeterReadingDate": "Reading date",
-  "pages.condo.meter.Service": "Service",
+  "pages.condo.meter.Resource": "Resource",
   "pages.condo.warning.modal.SaveLabel": "Save",
   "pages.condo.warning.modal.LeaveLabel": "Leave",
   "pages.condo.residents.PageTitle": "Residents",
