+++ apps/condo/lang/ru.json
@@ -359,7 +359,7 @@
   "pages.condo.meter.Place": "Место",
   "pages.condo.meter.MeterReading": "Показание",
   "pages.condo.meter.MeterReadingDate": "Дата снятия",
-  "pages.condo.meter.Service": "Услуга",
+  "pages.condo.meter.Resource": "Ресурс",
   "pages.condo.warning.modal.SaveLabel": "Сохранить",
   "pages.condo.warning.modal.LeaveLabel": "Не сохранять",
   "pages.condo.residents.PageTitle": "Жители",
@@ -449,7 +449,7 @@
   "pages.condo.meter.CanNotBeEarlierThanFirstVerification": "Дата следующей поверки не может быть раньше предыдущей",
   "pages.condo.meter.MeterNumber": "№ прибора учета",
   "pages.condo.meter.MeterPlace": "Место установки счетчика",
-  "pages.condo.meter.NumberOfMeter": "Прибор учета номер",
+  "pages.condo.meter.NumberOfMeter": "Прибор учета №",
   "pages.condo.meter.MeterWithSameNumberIsExist": "Прибор учета с таким номером уже существует в организации",
   "pages.condo.meter.TariffsNumber": "Количество тарифов",
   "pages.condo.meter.InstallationDate": "Дата установки",
